Seperate analysis functions into their own files
