+++ github_api/analysis/Vulnerability.py
@@ -1,0 +1,56 @@
+from lxml import etree
+
+from .extentions import Extentions
+
+
+def vulnerabilities_analysis(
+    vex,
+    extension: Extentions,
+    specification: str,
+    vulnerabilities: dict,
+    lacks_vulnerabilities: dict,
+) -> None:
+
+    if specification == "OpenVEX":
+        if "statements" in vex.keys():
+            vulnerabilities["OpenVEX"].append(len(vex["statements"]))
+            if len(vex["statements"]) == 0:
+                lacks_vulnerabilities["OpenVEX"] += 1
+        else:
+            lacks_vulnerabilities["OpenVEX"] += 1
+
+    elif specification == "CSAF":
+        if "vulnerabilities" in vex.keys():
+            vulnerabilities["CSAF"].append(len(vex["vulnerabilities"]))
+            if len(vex["vulnerabilities"]) == 0:
+                lacks_vulnerabilities["CSAF"] += 1
+        else:
+            lacks_vulnerabilities["CSAF"] += 1
+
+    elif specification == "CycloneDX":
+        if extension == Extentions.JSON:
+            if "vulnerabilities" in vex.keys():
+                vulnerabilities["CycloneDX"].append(len(vex["vulnerabilities"]))
+                if len(vex["vulnerabilities"]) == 0:
+                    lacks_vulnerabilities["CycloneDX"] += 1
+            else:
+                lacks_vulnerabilities["CycloneDX"] += 1
+        elif extension == Extentions.XML:
+            namespace = etree.QName(vex.tag).namespace
+            for vulns in vex.findall(etree.QName(namespace, "vulnerabilities")):
+                vulnerabilities["CycloneDX"].append(len(vulns))
+                if len(vulns) == 0:
+                    lacks_vulnerabilities["CycloneDX"] += 1
+                break
+
+    elif specification == "SPDX" and "@graph" in vex.keys():
+        vuln_count = 0
+        for element in vex["@graph"]:
+            if (
+                element["type"] == "security_Vulnerability"
+                or element["type"] == "Vulnerability"
+            ):
+                vuln_count += 1
+        vulnerabilities["SPDX"].append(vuln_count)
+        if vuln_count == 0:
+            lacks_vulnerabilities["SPDX"] += 1
