+++ github_api/analysis/extentions.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class Extentions(Enum):
+    JSON = 1
+    XML = 2
