+++ github_api/digest_vex.py
@@ -1,840 +1,20 @@
 import argparse
-import re
 from collections import defaultdict
 from copy import deepcopy
-from enum import Enum
 from statistics import mean, median, mode
 
 import jsonc  # Helps with parsing illegal Json
+from analysis.extentions import Extentions
+from analysis.Rating import ratings_analysis
+from analysis.Repository import repository_analysis
+from analysis.Specification import spesification_analysis
+from analysis.Status import status_analysis
+from analysis.Tools import tools_analysis
+from analysis.Vulnerability import vulnerabilities_analysis
+from analysis.Vulnerability_database import database_analysis
 from database import vexDB
 from lxml import etree
 from tqdm import tqdm
-
-# Regex constants
-ALPH = "a-zA-Z"
-NUM = "0-9"
-
-
-# Source - https://stackoverflow.com/a/68400507
-# Posted by Mark, modified by community. See post 'Timeline' for change history
-# Retrieved 2026-04-17, License - CC BY-SA 4.0
-
-
-# We could also do class Finger(IntEnum) its equivalent.
-class Status(int, Enum):
-    def __new__(cls, value, label):
-        # Initialise an instance of the Finger enum class
-        obj = int.__new__(cls, value)
-        # Calling print(type(obj)) returns <enum 'Finger'>
-        # If we don't set the _value_ in the Enum class, an error will be raised.
-        obj._value_ = value
-        # Here we add an attribute to the finger class on the fly.
-        # One may want to use setattr to be more explicit; note the python docs don't do this
-        obj.label = label
-        return obj
-
-    FIXED = (3, "FIXED")
-    AFFECTED = (2, "AFFECTED")
-    NOT_AFFECTED = (1, "NOT_AFFECTED")
-    UNDER_INVESTIGATION = (0, "UNDER_INVESTIGATION")
-    UNKNOWN = (-1, "UNKNOWN")
-
-    @classmethod
-    def from_str(cls, input_str):
-        for finger in cls:
-            if finger.label == input_str:
-                return finger
-        raise ValueError(f"{cls.__name__} has no value matching {input_str}")
-
-
-class Extentions(Enum):
-    JSON = 1
-    XML = 2
-
-# 4 Make plots
-
-
-def tools_analysis(
-    vex, extention: Extentions, specification: str, buckets: dict
-) -> dict:
-    """
-    Extracts the name of the tools used to generate the vex
-
-    Parameters
-    vex - the Vex file
-    extention - the extention of the vex file, this is so we can handle both json and xml
-    specification - the specification of the current Vex file
-    buckets - the datastructure we add another tool to
-
-    Returns
-    buckets
-    """
-    found_tool = False
-    if extention == Extentions.JSON:
-        if specification == "OpenVEX" and "tooling" in vex.keys():
-            buckets[specification][vex["tooling"]] += 1
-            found_tool = True
-
-        elif specification == "CSAF" and "document" in vex.keys():
-            # CSAF dosen't have a "tools" field, but a tool could be a publisher.
-            buckets[specification][vex["document"]["publisher"]["name"]] += 1
-            found_tool = True
-
-        elif (
-            specification == "CycloneDX"
-            and "metadata" in vex.keys()
-            and "tools"
-            in vex["metadata"].keys()  # There has to be something in the tools
-            and len(vex["metadata"]["tools"]) != 0
-        ):
-
-            # Handle different kind of tools
-            if type(vex["metadata"]["tools"]) == dict:
-                if "components" in vex["metadata"]["tools"].keys():
-                    tools = vex["metadata"]["tools"]["components"]
-                elif "services" in vex["metadata"]["tools"]:
-                    tools = vex["metadata"]["tools"]["services"]
-            elif type(vex["metadata"]["tools"]) == list:
-                tools = vex["metadata"]["tools"]
-
-            for tool in tools:
-                buckets[specification][tool["name"]] += 1
-                found_tool = True
-
-        elif specification == "SPDX" and "@graph" in vex.keys():
-            for entry in vex["@graph"]:
-                if entry["type"] == "CreationInfo" and "createdUsing" in entry.keys():
-                    for tool in entry["createdUsing"]:
-                        buckets[specification][tool] += 1
-                        found_tool = True
-
-    elif extention == Extentions.XML:
-        if specification == "CycloneDX":
-            namespace = etree.QName(vex.tag).namespace
-            for metadata in vex.findall(etree.QName(namespace, "metadata")):
-                for tools in metadata.findall(etree.QName(namespace, "tools")):
-                    for tool in tools:
-                        for sub_element in tool:
-                            if etree.QName(sub_element.tag).localname == "name":
-                                buckets[specification][sub_element.text] += 1
-                                found_tool = True
-    if found_tool:
-        buckets[specification]["count"] += 1
-    return buckets
-
-
-def spesification_analysis(
-    vex, extention: Extentions, specification: str, buckets: dict
-) -> dict:
-    """
-    Extracts the version number of the spesification
-
-    Parameters
-    vex - the Vex file
-    extention - the extention of the vex file, this is so we can handle both json and xml
-    specification - the specification of the current Vex file
-    buckets - the datastructure we add another version number to
-
-    Returns
-    buckets
-    """
-    found_spesification = False
-    if extention == Extentions.JSON:
-        if specification == "OpenVEX" and "@context" in vex.keys():
-            context = vex["@context"]
-            version = context.replace("https://openvex.dev/ns/v", "")
-            buckets[specification][version] += 1
-            found_spesification = True
-
-        elif specification == "CSAF" and "document" in vex.keys():
-            buckets[specification][vex["document"]["csaf_version"]] += 1
-            found_spesification = True
-
-        elif specification == "CycloneDX" and "specVersion" in vex.keys():
-
-            buckets[specification][vex["specVersion"]] += 1
-            found_spesification = True
-
-        elif specification == "SPDX" and "@graph" in vex.keys():
-            for entry in vex["@graph"]:
-                if entry["type"] == "CreationInfo" and "specVersion" in entry.keys():
-                    buckets[specification][entry["specVersion"]] += 1
-                    found_spesification = True
-
-    elif extention == Extentions.XML:
-        if specification == "CycloneDX":
-            namespace = etree.QName(vex.tag).namespace
-            version = namespace.replace("http://cyclonedx.org/schema/bom/", "")
-            buckets[specification][version] += 1
-            found_spesification = True
-
-    if found_spesification:
-        buckets[specification]["count"] += 1
-    return buckets
-
-
-def database_analysis(
-    vex, extention: Extentions, specification: str, buckets: dict
-) -> dict:
-    """
-    Extracts the name of the databases the vulnerabilites come from
-
-    Parameters
-    vex - the Vex file
-    extention - the extention of the vex file, this is so we can handle both json and xml
-    specification - the specification of the current Vex file
-    buckets - the datastructure we add the databases to
-
-    Returns
-    buckets
-    """
-    found_vulnerability_database = False
-    if extention == Extentions.JSON:
-        if specification == "OpenVEX" and "statements" in vex.keys():
-            for statement in vex["statements"]:
-                if type(statement) != dict:
-                    continue
-                if (
-                    "vulnerability" in statement.keys()
-                    and type(statement["vulnerability"]) == dict
-                    and "name" in statement["vulnerability"].keys()
-                ):
-                    buckets[
-                        strip_vulnarability_to_database(
-                            statement["vulnerability"]["name"]
-                        )
-                    ] += 1
-                    found_vulnerability_database = True
-                    if "aliases" in statement["vulnerability"].keys():
-                        for alias in statement["vulnerability"]["aliases"]:
-                            buckets[strip_vulnarability_to_database(alias)] += 1
-
-        elif specification == "CSAF" and "vulnerabilities" in vex.keys():
-            for vulnerability in vex["vulnerabilities"]:
-                if "cve" in vulnerability.keys():
-                    buckets[strip_vulnarability_to_database(vulnerability["cve"])] += 1
-                    found_vulnerability_database = True
-                if "ids" in vulnerability.keys():
-                    for id in vulnerability["ids"]:
-                        if type(id) != str:
-                            continue
-                        buckets[strip_vulnarability_to_database(id)] += 1
-                        found_vulnerability_database = True
-
-        elif specification == "CycloneDX" and "vulnerabilities" in vex.keys():
-            for vulnerability in vex["vulnerabilities"]:
-                if type(vulnerability) == dict and "id" in vulnerability.keys():
-                    buckets[strip_vulnarability_to_database(vulnerability["id"])] += 1
-                    found_vulnerability_database = True
-
-        elif specification == "SPDX" and "@graph" in vex.keys():
-            for entry in vex["@graph"]:
-                if (
-                    entry["type"] == "Vulnerability"
-                    and "externalIdentifier" in entry.keys()
-                ):
-                    for external_identifier in entry["externalIdentifier"]:
-                        if external_identifier["type"] == "ExternalIdentifier" and (
-                            external_identifier["externalIdentifierType"] == "cve"
-                            or external_identifier["externalIdentifierType"]
-                            == "securityOther"
-                        ):
-                            buckets[
-                                strip_vulnarability_to_database(
-                                    external_identifier["identifier"]
-                                )
-                            ] += 1
-                            found_vulnerability_database = True
-
-    elif extention == Extentions.XML:
-        if specification == "CycloneDX":
-            namespaces_keys = list(vex.nsmap.keys())
-            for key in namespaces_keys:
-                for vulnerabilities in vex.findall(
-                    f"{f"{key}:" if key else ""}vulnerabilities", namespaces=vex.nsmap
-                ):
-                    for vulnerability in vulnerabilities.findall(
-                        f"{f"{key}:" if key else ""}vulnerability", namespaces=vex.nsmap
-                    ):
-                        for id in vulnerability.findall(
-                            f"{f"{key}:" if key else ""}id", namespaces=vex.nsmap
-                        ):
-                            buckets[strip_vulnarability_to_database(id.text)] += 1
-                            found_vulnerability_database = True
-
-    if found_vulnerability_database:
-        buckets["count"] += 1
-    return buckets
-
-
-def strip_vulnarability_to_database(input: str) -> str:
-    """Tries to extract the datbase from the input"""
-    if input == None or input.lower() == "none":
-        return "NULL"
-
-    first_dash = input.rfind("-")
-    if first_dash == -1:
-        return "INVALID"
-
-    if input.startswith("http"):
-        vulnerability = link_sanitation(input)
-    else:
-        vulnerability = input
-
-    if vulnerability.find(":") != -1:
-        first_half = vulnerability.split(":")[0]
-        database = first_half[: first_half.rfind("-")]
-    else:
-        identifier_regex = f"[{ALPH}]{{2,7}}"
-        min_one_number = f"(-(?=[{ALPH}]*[{NUM}])"
-        segment_regex = f"({min_one_number}([{ALPH}{NUM}]){{2,15}}){{1,3}})"
-        database_regex = f"{identifier_regex}{segment_regex}"
-        # [a-zA-Z]{2,7}((-(?=[a-zA-Z]*[0-9])([a-zA-Z0-9]){2,15}){1,3})
-
-        result = re.match(pattern=database_regex, string=vulnerability)
-        if not result:
-            return "NOT_DATABASE"
-        else:
-            dash = vulnerability.find("-")
-            first_half = vulnerability[:dash]
-
-            ident = re.match(pattern=identifier_regex, string=first_half)
-            database = ident.group(0)
-
-    return database
-
-
-def link_sanitation(vulnerability: str) -> str:
-    """Removes the link from the vulnerablility"""
-    seperators = [
-        "/",
-        ",",
-        "=",
-        "?",
-        ":",
-    ]
-    counter = 0
-    for char in reversed(vulnerability):
-        if char in seperators:
-            break
-        else:
-            counter += 1
-
-    start_index = len(vulnerability) - counter
-    return vulnerability[start_index:]
-
-
-def vulnerabilities_analysis(
-    vex,
-    extension: Extentions,
-    specification: str,
-    vulnerabilities: dict,
-    lacks_vulnerabilities: dict,
-) -> None:
-
-    if specification == "OpenVEX":
-        if "statements" in vex.keys():
-            vulnerabilities["OpenVEX"].append(len(vex["statements"]))
-            if len(vex["statements"]) == 0:
-                lacks_vulnerabilities["OpenVEX"] += 1
-        else:
-            lacks_vulnerabilities["OpenVEX"] += 1
-
-    elif specification == "CSAF":
-        if "vulnerabilities" in vex.keys():
-            vulnerabilities["CSAF"].append(len(vex["vulnerabilities"]))
-            if len(vex["vulnerabilities"]) == 0:
-                lacks_vulnerabilities["CSAF"] += 1
-        else:
-            lacks_vulnerabilities["CSAF"] += 1
-
-    elif specification == "CycloneDX":
-        if extension == Extentions.JSON:
-            if "vulnerabilities" in vex.keys():
-                vulnerabilities["CycloneDX"].append(len(vex["vulnerabilities"]))
-                if len(vex["vulnerabilities"]) == 0:
-                    lacks_vulnerabilities["CycloneDX"] += 1
-            else:
-                lacks_vulnerabilities["CycloneDX"] += 1
-        elif extension == Extentions.XML:
-            namespace = etree.QName(vex.tag).namespace
-            for vulns in vex.findall(etree.QName(namespace, "vulnerabilities")):
-                vulnerabilities["CycloneDX"].append(len(vulns))
-                if len(vulns) == 0:
-                    lacks_vulnerabilities["CycloneDX"] += 1
-                break
-
-    elif specification == "SPDX" and "@graph" in vex.keys():
-        vuln_count = 0
-        for element in vex["@graph"]:
-            if (
-                element["type"] == "security_Vulnerability"
-                or element["type"] == "Vulnerability"
-            ):
-                vuln_count += 1
-        vulnerabilities["SPDX"].append(vuln_count)
-        if vuln_count == 0:
-            lacks_vulnerabilities["SPDX"] += 1
-
-
-def status_analysis(
-    vex, extention: Extentions, specification: str, buckets: dict
-) -> dict:
-    """
-    Extracts the status of the vulnerability and count the occurences
-
-    Parameters
-    vex - the Vex file
-    extention - the extention of the vex file, this is so we can handle both json and xml
-    specification - the specification of the current Vex file
-    buckets - the datastructure we add the databases to
-
-    Returns
-    buckets
-    """
-    found_status = False
-    if extention == Extentions.JSON:
-        if specification == "OpenVEX" and "statements" in vex.keys():
-            for statement in vex["statements"]:
-                if type(statement) != dict:
-                    continue
-                if "status" in statement.keys():
-                    match (statement["status"]):
-                        case "not_affected":
-                            buckets[specification][Status.NOT_AFFECTED.label] += 1
-                            found_status = True
-                            break
-                        case "affected":
-                            buckets[specification][Status.AFFECTED.label] += 1
-                            found_status = True
-                            break
-                        case "fixed":
-                            buckets[specification][Status.FIXED.label] += 1
-                            found_status = True
-                            break
-                        case "under_investigation":
-                            buckets[specification][
-                                Status.UNDER_INVESTIGATION.label
-                            ] += 1
-                            found_status = True
-                            break
-                        case _:
-                            buckets[specification][Status.UNKNOWN.label] += 1
-                            found_status = True
-                            break
-
-        elif specification == "CSAF" and "vulnerabilities" in vex.keys():
-            for vulnerability in vex["vulnerabilities"]:
-                if "product_status" in vulnerability.keys():
-                    final_products_properties = defaultdict()
-                    for property, products in vulnerability["product_status"].items():
-                        for product in products:
-                            if product not in final_products_properties.keys():
-                                final_products_properties[product] = Status.UNKNOWN
-                            match property:
-                                case "first_affected":
-                                    if (
-                                        final_products_properties[product]
-                                        < Status.AFFECTED
-                                    ):
-                                        final_products_properties[product] = (
-                                            Status.AFFECTED
-                                        )
-                                    break
-                                case "first_fixed":
-                                    if (
-                                        final_products_properties[product]
-                                        < Status.FIXED
-                                    ):
-                                        final_products_properties[product] = (
-                                            Status.FIXED
-                                        )
-                                    break
-                                case "fixed":
-                                    if (
-                                        final_products_properties[product]
-                                        < Status.FIXED
-                                    ):
-                                        final_products_properties[product] = (
-                                            Status.FIXED
-                                        )
-                                    break
-                                case "known_affected":
-                                    if (
-                                        final_products_properties[product]
-                                        < Status.AFFECTED
-                                    ):
-                                        final_products_properties[product] = (
-                                            Status.AFFECTED
-                                        )
-                                    break
-                                case "known_not_affected":
-                                    if (
-                                        final_products_properties[product]
-                                        < Status.NOT_AFFECTED
-                                    ):
-                                        final_products_properties[product] = (
-                                            Status.NOT_AFFECTED
-                                        )
-                                    break
-                                case "last_affected":
-                                    if (
-                                        final_products_properties[product]
-                                        < Status.AFFECTED
-                                    ):
-                                        final_products_properties[product] = (
-                                            Status.AFFECTED
-                                        )
-                                    break
-                                case "recommended":
-                                    if (
-                                        final_products_properties[product]
-                                        < Status.FIXED
-                                    ):
-                                        final_products_properties[product] = (
-                                            Status.FIXED
-                                        )
-                                    break
-                                case "under_investigation":
-                                    if (
-                                        final_products_properties[product]
-                                        < Status.UNDER_INVESTIGATION
-                                    ):
-                                        final_products_properties[product] = (
-                                            Status.UNDER_INVESTIGATION
-                                        )
-                                    break
-                        for product, affects in final_products_properties.items():
-                            found_status = True
-                            buckets[specification][affects.label] += 1
-
-        elif specification == "CycloneDX" and "vulnerabilities" in vex.keys():
-            for vulnerability in vex["vulnerabilities"]:
-                if (
-                    type(vulnerability) == dict
-                    and "id" in vulnerability.keys()
-                    and "affects" in vulnerability.keys()
-                ):
-                    for affected in vulnerability["affects"]:
-                        found_status = True
-                        if "versions" in affected.keys():
-                            for version in affected["versions"]:
-                                if type(version) == dict and "status" in version.keys():
-                                    match version["status"]:
-                                        case "affected":
-                                            buckets[specification][
-                                                Status.AFFECTED.label
-                                            ] += 1
-                                        case "unaffected":
-                                            buckets[specification][
-                                                Status.NOT_AFFECTED.label
-                                            ] += 1
-                                        case "unknown":
-                                            buckets[specification][
-                                                Status.UNKNOWN.label
-                                            ] += 1
-                        else:
-                            buckets[specification][Status.UNKNOWN.label] += 1
-
-        elif specification == "SPDX" and "@graph" in vex.keys():
-            relationships = [
-                "VexAffectedVulnAssessmentRelationship",
-                "VexFixedVulnAssessmentRelationship",
-                "VexNotAffectedVulnAssessmentRelationship",
-                "VexUnderInvestigationVulnAssessmentRelationship",
-            ]
-            for entry in vex["@graph"]:
-                if (
-                    entry["type"] == "Relationship"
-                    and entry["relationshipType"] in relationships
-                ):
-                    found_status = True
-                    match entry["relationshipType"]:
-                        case "VexAffectedVulnAssessmentRelationship":
-                            buckets[specification][Status.AFFECTED.label] += 1
-                            break
-                        case "VexFixedVulnAssessmentRelationship":
-                            buckets[specification][Status.FIXED.label] += 1
-                            break
-                        case "VexNotAffectedVulnAssessmentRelationship":
-                            buckets[specification][Status.NOT_AFFECTED.label] += 1
-                            break
-                        case "VexUnderInvestigationVulnAssessmentRelationship":
-                            buckets[specification][Status.NOT_AFFECTED.label] += 1
-                            break
-
-    elif extention == Extentions.XML:
-        if specification == "CycloneDX":
-            namespaces_keys = list(vex.nsmap.keys())
-            for key in namespaces_keys:
-                for vulnerabilities in vex.findall(
-                    f"{f"{key}:" if key else ""}vulnerabilities", namespaces=vex.nsmap
-                ):
-                    for vulnerability in vulnerabilities.findall(
-                        f"{f"{key}:" if key else ""}vulnerability", namespaces=vex.nsmap
-                    ):
-                        for affects in vulnerability.findall(
-                            f"{f"{key}:" if key else ""}affects", namespaces=vex.nsmap
-                        ):
-                            for target in affects.findall(
-                                f"{f"{key}:" if key else ""}target",
-                                namespaces=vex.nsmap,
-                            ):
-                                for versions in target.findall(
-                                    f"{f"{key}:" if key else ""}versions",
-                                    namespaces=vex.nsmap,
-                                ):
-                                    for version in versions.findall(
-                                        f"{f"{key}:" if key else ""}version",
-                                        namespaces=vex.nsmap,
-                                    ):
-                                        for status in version.findall(
-                                            f"{f"{key}:" if key else ""}status",
-                                            namespaces=vex.nsmap,
-                                        ):
-                                            found_status = True
-                                            match status.text:
-                                                case "affected":
-                                                    buckets[specification][
-                                                        Status.AFFECTED.label
-                                                    ] += 1
-                                                case "unaffected":
-                                                    buckets[specification][
-                                                        Status.NOT_AFFECTED.label
-                                                    ] += 1
-                                                case "unknown":
-                                                    buckets[specification][
-                                                        Status.UNKNOWN.label
-                                                    ] += 1
-
-    if found_status:
-        buckets[specification]["count"] += 1
-    return buckets
-
-
-def ratings_analysis(
-    vex, extention: Extentions, specification: str, buckets: dict
-) -> dict:
-    """
-    Extracts the severity of vulnerabilities
-
-    Parameters
-    vex - the Vex file
-    extention - the extention of the vex file, this is so we can handle both json and xml
-    specification - the specification of the current Vex file
-    buckets - the datastructure we add the databases to
-
-    Returns
-    buckets
-    """
-    found_rating = False
-    if extention == Extentions.JSON:
-        if specification == "CSAF" and "vulnerabilities" in vex.keys():
-            for vulnerability in vex["vulnerabilities"]:
-                if "scores" in vulnerability.keys():
-                    for item in vulnerability["scores"]:
-                        for system in item.keys():
-                            match system:
-                                case "cvss_v2":
-                                    if item[system]["baseScore"]:
-                                        found_rating = True
-                                        buckets[specification]["CVSS"]["2"] += 1
-                                        score = float(item[system]["baseScore"])
-                                        buckets[specification]["ratings"].append(score)
-                                    break
-                                case "cvss_v3":
-                                    if item[system]["baseScore"]:
-                                        found_rating = True
-                                        buckets[specification]["CVSS"]["3"] += 1
-                                        score = float(item[system]["baseScore"])
-                                        buckets[specification]["ratings"].append(score)
-                                    break
-                                case _:
-                                    break
-
-        elif specification == "CycloneDX" and "vulnerabilities" in vex.keys():
-            for vulnerability in vex["vulnerabilities"]:
-                if type(vulnerability) == dict and "ratings" in vulnerability.keys():
-                    for item in vulnerability["ratings"]:
-                        if "method" in item.keys():
-                            match item["method"]:
-                                case "CVSSv2":
-                                    if "score" in item.keys():
-                                        score = item["score"]
-                                        found_rating = True
-                                        buckets[specification]["CVSS"]["2"] += 1
-                                        buckets[specification]["ratings"].append(score)
-                                    break
-                                case "CVSSv3" | "CVSSv31":
-                                    if "score" in item.keys():
-                                        score = item["score"]
-                                        found_rating = True
-                                        buckets[specification]["CVSS"]["3"] += 1
-                                        buckets[specification]["ratings"].append(score)
-                                    break
-                                case "other":
-                                    if (
-                                        "score" in item.keys()
-                                        and "source" in item.keys()
-                                    ):
-                                        found_rating = True
-                                        score = item["score"]
-                                        name = item["source"]["name"]
-                                        buckets[specification]["other"].append(
-                                            {"score": score, "name": name}
-                                        )
-                                        buckets[specification]["ratings"].append(score)
-                                    break
-                                # While supported none were found in the dataset
-                                # so the code has been left incomplete
-                                case "CVSSv4":
-                                    if "score" in item.keys():
-                                        score = item["score"]
-                                        found_rating = True
-                                        buckets[specification]["CVSS"]["4"] += 1
-                                        buckets[specification]["ratings"].append(score)
-                                case "OWASP":
-                                    break
-
-        elif specification == "SPDX" and "@graph" in vex.keys():
-            relationships = [
-                "CvssV2VulnAssessmentRelationship",
-                "CvssV3VulnAssessmentRelationship",
-                "CvssV4VulnAssessmentRelationship",
-                "EpssVulnAssessmentRelationship",
-                "SsvcVulnAssessmentRelationship",
-            ]
-            for entry in vex["@graph"]:
-                if (
-                    entry["type"] == "Relationship"
-                    and entry["relationshipType"] in relationships
-                    and "security_score" in entry
-                ):
-                    found_rating = True
-                    match entry["relationshipType"]:
-                        case "CvssV2VulnAssessmentRelationship":
-                            buckets[specification]["CVSS"]["2"] += 1
-                            buckets[specification]["ratings"].append(
-                                float(entry["security_score"])
-                            )
-                            break
-                        case "CvssV3VulnAssessmentRelationship":
-                            buckets[specification]["CVSS"]["3"] += 1
-                            buckets[specification]["ratings"].append(
-                                float(entry["security_score"])
-                            )
-                            break
-                        case "CvssV4VulnAssessmentRelationship":
-                            buckets[specification]["CVSS"]["4"] += 1
-                            buckets[specification]["ratings"].append(
-                                float(entry["security_score"])
-                            )
-                            break
-
-    elif extention == Extentions.XML:
-        if specification == "CycloneDX":
-            namespaces_keys = list(vex.nsmap.keys())
-            for key in namespaces_keys:
-                for vulnerabilities in vex.findall(
-                    f"{f"{key}:" if key else ""}vulnerabilities", namespaces=vex.nsmap
-                ):
-                    for vulnerability in vulnerabilities.findall(
-                        f"{f"{key}:" if key else ""}vulnerability", namespaces=vex.nsmap
-                    ):
-                        for ratings in vulnerability.findall(
-                            f"{f"{key}:" if key else ""}ratings", namespaces=vex.nsmap
-                        ):
-                            for rating in ratings.findall(
-                                f"{f"{key}:" if key else ""}rating",
-                                namespaces=vex.nsmap,
-                            ):
-                                score_str = rating.find(
-                                    f"{f"{key}:" if key else ""}score",
-                                    namespaces=vex.nsmap,
-                                )
-                                method = rating.find(
-                                    f"{f"{key}:" if key else ""}method",
-                                    namespaces=vex.nsmap,
-                                )
-
-                                try:
-                                    score = float(score_str.text)
-                                except Exception as err:
-                                    continue
-                                if method is not None:
-                                    match method.text:
-                                        case "CVSSv2":
-                                            found_rating = True
-                                            buckets[specification]["CVSS"]["2"] += 1
-                                            buckets[specification]["ratings"].append(
-                                                score
-                                            )
-                                            break
-                                        case "CVSSv3" | "CVSSv31":
-                                            found_rating = True
-                                            buckets[specification]["CVSS"]["3"] += 1
-                                            buckets[specification]["ratings"].append(
-                                                score
-                                            )
-                                            break
-                                        case "other":
-                                            source = rating.find(
-                                                f"{f"{key}:" if key else ""}source",
-                                                namespaces=vex.nsmap,
-                                            )
-                                            if source is None:
-                                                continue
-                                            name_element = source.find(
-                                                f"{f"{key}:" if key else ""}name",
-                                                namespaces=vex.nsmap,
-                                            )
-                                            if name_element is None:
-                                                continue
-                                            found_rating = True
-                                            name = source.find(
-                                                f"{f"{key}:" if key else ""}name",
-                                                namespaces=vex.nsmap,
-                                            ).text
-                                            buckets[specification]["other"].append(
-                                                {"score": score, "name": name}
-                                            )
-                                            buckets[specification]["ratings"].append(
-                                                score
-                                            )
-                                            break
-                                        case "CVSSv4":
-                                            found_rating = True
-                                            buckets[specification]["CVSS"]["4"] += 1
-                                            buckets[specification]["ratings"].append(
-                                                score
-                                            )
-                                            break
-                                        # While supported none were found in the dataset
-                                        # so the code has been left incomplete
-                                        case "OWASP":
-                                            break
-    if found_rating:
-        buckets[specification]["count"] += 1
-    return buckets
-
-
-def repository_analysis(document, specification: str, buckets: dict) -> dict:
-    """
-    Extracts the repository the file comes from
-
-    Parameters
-    document - the mongodb document
-    specification - the specification of the current document file
-    buckets - the datastructure we add the repository data to
-
-    Returns
-    buckets
-    """
-    commit_url = document["commit_url"]
-    repo = commit_url.split("repos/")
-    repo = repo[1].split("/", 2)[0] + "/" + repo[1].split("/", 2)[1]
-    buckets[specification][repo] += 1
-    return buckets
 
 
 def input_arguments() -> argparse.Namespace:
